sql_backend/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
google-generativeai==0.8.2
PyMySQL==1.1.1
requests>=2.31.0
SQLAlchemy==2.0.35
PyInputPlus==0.2.12
gunicorn==21.2.0
cryptography>=40.0.0
python-dotenv==1.0.1
Flask-PyMongo==2.3.0
Flask-Mail==0.10.0
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.1
Werkzeug==3.0.3